feat: category

+++ database/migrations/2026_07_07_005330_add_tags_to_news_table.php
@@ -8,15 +8,19 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('news', function (Blueprint $table) {
-            $table->string('tags')->nullable()->after('links');
-        });
+        if (! Schema::hasColumn('news', 'tags')) {
+            Schema::table('news', function (Blueprint $table) {
+                $table->string('tags')->nullable()->after('links');
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::table('news', function (Blueprint $table) {
-            $table->dropColumn('tags');
-        });
+        if (Schema::hasColumn('news', 'tags')) {
+            Schema::table('news', function (Blueprint $table) {
+                $table->dropColumn('tags');
+            });
+        }
     }
 };
